requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene==2.0.0b5
webdriver-manager==3.7.0
python-dotenv==1.0.0